Audit Final: Echo resource, NPC precision, wound logic hardening, and seasonal regularity

+++ app/src/main/java/com/grimreich/core/GameState.kt
@@ -459,7 +459,7 @@ data class NpcDto(
     val factionId: String?,
     val personality: String,
     val startNodeId: String?,
-    val stability: Int
+    val stability: Float
 )
 
 @Serializable
